backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
yfinance
httpx
pytest
pytest-asyncio
pytest-mock
pytest-html
python-dotenv
aiosqlite
python-jose[cryptography]
